Migrated keygen history transaction sending to a dedicated module.
